% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_col_names.R
\name{clean_col_names}
\alias{clean_col_names}
\title{Clean column names by converting case and replacing non-standard characters to _}
\usage{
clean_col_names(data, case = c("lower", "upper"))
}
\arguments{
\item{data}{dataframe to clean up column names}

\item{case}{string for lower or upper case}
}
\value{
the dataframe with updated column names
}
\description{
Clean column names by converting case and replacing non-standard characters to _
}
\examples{
df <- toolkitr::profile_example \%>\% clean_col_names()
}
