% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decile.R
\name{decile}
\alias{decile}
\title{Calculate percentile by volume}
\usage{
decile(
  data,
  unique_id,
  value_var,
  num_groups,
  new_col,
  tie_breaker_var = {
     {
         unique_id
     }
 },
  verbose = TRUE,
  calc_type = "group by value"
)
}
\arguments{
\item{data}{data frame with value column}

\item{value_var}{variable to denote the numerical variable}

\item{num_groups}{number of groups, e.g. 10 means decile plus a 0 group}

\item{tie_breaker_var}{variable to denote what to use for tiebreakers}

\item{verbose}{verbose option of readout}

\item{calc_type}{2 options "equal sum" or "group by value". Former uses continuous breaks and the latter groups observations based on their value as well. i.e. if 2 customers have the value 5, they will not be split across groups and will be forced together. The ones in a lower group will be promoted up}

\item{group_label}{string to name the deciles}
}
\value{
Returns the same dataframe with an additional decile column
}
\description{
Create percentiles for a single column. This appends the decile column to the dataset.
Will also keep the same volume in the same category
}
\details{
NAs and negatives are considered 0.
}
\examples{
decile_df <- toolkitr::profile_example \%>\% dplyr::group_by(market_trx_post_decile) \%>\%
decile(unique_id = "source_id", value_var = "market_decile_nbrx_pre",
tie_breaker_var = "source_id", num_groups = 5, new_col = "quintile")

decile_df <- toolkitr::profile_example \%>\% decile(unique_id = "source_id", value_var = "market_decile_nbrx_pre",
tie_breaker_var = "source_id", num_groups = 5,
new_col = "quintile", calc_type = "group by value")
}
