% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_conc_curves.R
\name{draw_conc_curves}
\alias{draw_conc_curves}
\title{Title}
\usage{
draw_conc_curves(
  data,
  value_var,
  value_breaks = seq(0, 1, by = 0.05),
  cust_breaks = c(0.2, 0.5),
  keep_first_n_points = 14
)
}
\arguments{
\item{data}{data frame with value column}

\item{value_var}{variable to denote the numerical variable}

\item{value_breaks}{sequence of value \%s to flag for graphing}

\item{cust_breaks}{set of customer \%s to flag for graphing}

\item{keep_first_n_points}{how many of the first n-points to keep for graphing the steep slope}
}
\value{
Returns a new dataframe of all the customers x groups and concentration calculations
}
\description{
Title
}
\examples{
df_curves <- toolkitr::profile_example \%>\% draw_conc_curves(value_var = "market_decile_nbrx_pre")
}
